Move history watcher to watcher directory
