+++ client/src/components/History/adapters/HistoryPanelProxy.js
@@ -4,7 +4,7 @@
  */
 import Backbone from "backbone";
 import { createDatasetCollection } from "components/History/model/queries";
-import { startWatchingHistory } from "store/historyStore/model/watchHistory";
+import { startWatchingHistory } from "watcher/watchHistory";
 import { useHistoryItemsStore } from "stores/historyItemsStore";
 import { useHistoryStore } from "stores/historyStore";
